connect resolver and validation pipelines
